More precise timing in scene sequence

+++ utility.hpp
@@ -43,7 +43,7 @@ static inline QString MakeScreenRegionSwitchName(
 }
 
 static inline QString MakeSceneRoundTripSwitchName(
-	const QString& scene1, const QString& scene2, const QString& transition, int delay)
+	const QString& scene1, const QString& scene2, const QString& transition, double delay)
 {
 	return scene1 + QStringLiteral(" -> wait for ") + QString::number(delay)
 		+ QStringLiteral(" seconds -> ") + scene2 + QStringLiteral(" (using ") + transition
